Cert 2 of 5 obtained!
Good stretching of the brain for the last exercise in data structures, very quickly knew not to double-loop. Also realized that I could convert a list entry to a string... that I shouldn't... and that I could just go off of a sub for a piece that I need. Also made a fun script that emphasizes what the ongoing learning journey looks like.